more edits

+++ src/main/java/com/advert/service/UserService.java
@@ -1,9 +1,11 @@
 package com.advert.service;
 
+import com.advert.handler.InvalidDataException;
 import com.advert.model.UserDto;
 import com.advert.repository.UserRepository;
 import com.advert.repository.model.UserEntity;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 @Service
@@ -12,16 +14,21 @@ public class UserService {
     private final UserRepository userRepository;
 
     public void createNewUser(UserDto userDto){
-        userRepository.save(new UserEntity(null, userDto.getUsername(), userDto.getEmail(), userDto.getPassword(), userDto.getPhone(), false));
+        BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
+        userRepository.save(new UserEntity(null, userDto.getUsername(), userDto.getEmail(), bCryptPasswordEncoder.encode(userDto.getPassword()), userDto.getPhone(), false));
     }
 
     public void updateUser(String username, UserDto userDto){
-        Long userId = userRepository.findByUsername(username).getUserID();
-        userRepository.save(new UserEntity(userId, userDto.getUsername(), userDto.getEmail(), userDto.getPassword(), userDto.getPhone(), false));
+        BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
+        Long userId = userRepository.findByUsername(username).get().getUserID();
+        userRepository.save(new UserEntity(userId, userDto.getUsername(), userDto.getEmail(), bCryptPasswordEncoder.encode(userDto.getPassword()), userDto.getPhone(), false));
     }
 
-    public void deleteUser(UserDto userDto){
-        userRepository.deleteByUsername(userDto.getUsername());
-
+    public void deleteUser(String username, UserDto userDto){
+        if(username.equals(userDto.getUsername())) {
+            userRepository.deleteByUsername(userDto.getUsername());
+        } else {
+            throw new InvalidDataException();
+        }
     }
 }
